Refactored all Request and Response Objects to be isolated into own packages
