+++ src/main/java/org/secureauth/sarestapi/data/Requests/AuthRequest.java
@@ -1,4 +1,4 @@
-package org.secureauth.sarestapi.data;
+package org.secureauth.sarestapi.data.Requests;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import org.secureauth.sarestapi.util.JSONUtil;
@@ -8,7 +8,6 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 /**
  * @author rrowcliffe@secureauth.com
- *
  *
 Copyright (c) 2015, SecureAuth
 All rights reserved.
@@ -24,15 +23,15 @@ Redistribution and use in source and binary forms, with or without modification,
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
 IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
 OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-**/
-
+ */
 @XmlRootElement
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class DFPConfirmRequest {
+public class AuthRequest {
 
-    private String user_id;
-    private String fingerprint_id;
-
+    String user_id;
+    private String type;
+    private String token;
+    private String factor_id;
 
 
     public String getUser_id() {
@@ -43,16 +42,33 @@ public class DFPConfirmRequest {
         this.user_id = user_id;
     }
 
-    public String getFingerprint_id() {
-        return fingerprint_id;
+    public String getType() {
+        return type;
     }
 
-    public void setFingerprint_id(String fingerprint_id) {
-        this.fingerprint_id = fingerprint_id;
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public String getFactor_id() {
+        return factor_id;
+    }
+
+    public void setFactor_id(String factor_id) {
+        this.factor_id = factor_id;
     }
 
     @Override
     public String toString(){
         return JSONUtil.convertObjectToJSON(this);
     }
+
 }
