+++ src/main/java/org/secureauth/sarestapi/data/Requests/PushToAcceptRequest.java
@@ -1,6 +1,7 @@
-package org.secureauth.sarestapi.data;
+package org.secureauth.sarestapi.data.Requests;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import org.secureauth.sarestapi.data.PushAcceptDetails;
 import org.secureauth.sarestapi.util.JSONUtil;
 
 import javax.xml.bind.annotation.XmlRootElement;
