+++ src/main/java/org/secureauth/sarestapi/interfaces/AuthenticationInterface.java
@@ -1,8 +1,8 @@
 package org.secureauth.sarestapi.interfaces;
 
 import org.secureauth.sarestapi.SAAccess;
-import org.secureauth.sarestapi.data.FactorsResponse;
-import org.secureauth.sarestapi.data.ResponseObject;
+import org.secureauth.sarestapi.data.Response.FactorsResponse;
+import org.secureauth.sarestapi.data.Response.ResponseObject;
 
 /**
  * Created by rrowcliffe on 5/6/16.
