+++ src/main/java/org/secureauth/sarestapi/interfaces/JavascriptResource.java
@@ -1,7 +1,7 @@
 package org.secureauth.sarestapi.interfaces;
 
 import org.secureauth.sarestapi.SAAccess;
-import org.secureauth.sarestapi.data.JSObjectResponse;
+import org.secureauth.sarestapi.data.Response.JSObjectResponse;
 
 /**
  * Created by rrowcliffe on 5/7/16.
